AnimalStuff/test/animalstuff/DuckTest.java
package animalstuff;


import org.junit.Test;
import static org.junit.Assert.*;

/**
 * DuckTest class for testing Duck class constructor and methods
 * 
 * @author Martin Dwyer
 * 
 * Prepared for ITEC 6130, Walden University, March 15, 2019
 */
public class DuckTest {
    
    public DuckTest() {
    }
    /**
     * Test of movement method
     */
    @Test
    public void testMovement() {
        //  Testing movement() method
        Duck instance = new Duck();
        String fastMovement = "I fly";
        String slowMovement = "I float, swim, or waddle on two legs";
        assertEquals("movement() method fails testing for fast movement", fastMovement, instance.movement(true));
        assertEquals("movement() method fails testing for slow movement", slowMovement, instance.movement(false));
        
    }
    
    /**
     * Test of sound method
     */
    @Test
    public void testSound() {
        // Testing sound() method
        Duck instance = new Duck();
        String expResult = "Quack quack quack";
        String result = instance.sound();
        assertEquals("sound() method fails testing", expResult, result);
        
    }

    /**
     * Test of equals method
     */
    @Test
    public void testEquals() {
        Object obj = null;
        Duck instance = new Duck();

        // Testing not equal to null object
        boolean testResult = instance.equals(obj);
        assertFalse("equals() fails null test", testResult);

        // Testing for known equal objects...
        Duck instanceEqual = new Duck();
        testResult = instance.equals(instanceEqual);
        assertTrue("equals() fails for known equal object...", testResult);

        // Testing for known not-equal objects...
        Bird instanceNotEqual = new Bird("Canary");
        testResult = instance.equals(instanceNotEqual);
        assertFalse("equals() fails for known equal object...", testResult);

        // Testing for impersonator: new Animal(Finch,Feathers)
        Animal faker = new Animal("Duck", "Feathers");
        testResult = instance.equals(faker);
        assertFalse("equals() fails the impersonator test", testResult);

    }
 }
    

